lab4/frontend/requirements.txt
flask==2.3.3
pika==1.3.2
psycopg2-binary==2.9.9
gunicorn==20.1.0
Pillow==10.2.0
